feat: 重构创建作业页面
